OAUTH2-207 Extract BaseCorsFilter class
