+++ modules/apps/oauth2-provider/oauth2-provider-rest/src/main/java/com/liferay/oauth2/provider/rest/internal/cors/servlet/filters/OAuth2CORSServletBaseFilter.java
@@ -1,0 +1,48 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.oauth2.provider.rest.internal.cors.servlet.filters;
+
+import com.liferay.oauth2.provider.rest.internal.cors.CORSSupport;
+import com.liferay.portal.kernel.servlet.BaseFilter;
+import com.liferay.portal.kernel.util.Validator;
+
+import java.util.Map;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+/**
+ * @author Carlos Sierra Andrés
+ */
+public abstract class OAuth2CORSServletBaseFilter extends BaseFilter {
+
+	@Override
+	public boolean isFilterEnabled(
+		HttpServletRequest request, HttpServletResponse response) {
+
+		if (Validator.isBlank(request.getHeader("Origin"))) {
+			return false;
+		}
+
+		return super.isFilterEnabled(request, response);
+	}
+
+	public void setHeaders(Map<String, String> headers) {
+		corsSupport.setHeaders(headers);
+	}
+
+	protected final CORSSupport corsSupport = new CORSSupport();
+
+}
